+++ TODO.txt
@@ -1,0 +1,10 @@
+* import mit hashprüfung für skip
+* integration load tracks mit proximity calc
+* passwort für Db von umgebungsvariable
+* commandlinetools für
+    tracks listen
+    tracks + alle artefakte löschen
+* punkte aus umkreis von punkt löschen / filtern
+
+* check all for vacuum and correct indexes
+* nur bei debug module und zeilennummern
